Implement the contextmanager interface, and use it during tests
If the user interrupts execution during a run -- or something bad
happens -- finish() is never called. Some bars suppress the caret in
the console, and only re-enable it in finish(), leaving it invisible
in these cases.

Python has an interface for addressing these needs: context managers.
So we turn bars into one and use it during the test.